Fire-Bot/requirements.txt
flask
google.generativeai
cryptography.fernet
psutil
pynput
matplotlib.pyplot
scapy
winotisy
pillow opencv-python
pytesseract